GatchaLife-backend/requirements.txt
asgiref==3.10.0
certifi==2025.11.12
charset-normalizer==3.4.4
Django==5.2.8
django-cors-headers==4.9.0
django-filter==25.2
djangorestframework==3.16.1
drf-yasg==1.21.11
idna==3.11
inflection==0.5.1
packaging==25.0
pillow==12.0.0
pytz==2025.2
PyYAML==6.0.3
requests==2.32.5
sqlparse==0.5.3
uritemplate==4.2.0
urllib3==2.5.0
psycopg2-binary==2.9.10
gunicorn==21.2.0
whitenoise==6.6.0
structlog==24.1.0
